extract connection logic to core
